requirements.txt
python-telegram-bot
python-dotenv
httpx
requests